Improve RGB→256-colour palette interpolation by better maths
Improve the way 24-bit colours are mapped to 8-bit ANSI palette by
observing that neither grayscale ramp nor 6×6×6 colour cube have
linear staps throughout.

For example, sequence of values in the cube is: 0, 95, 135, 175, 215
and 255.  The first step is 95 colours while the rest are only 40
colours.  Similarly, grayscale ramp (plus black and white) create
a sequence 0, 8, 18, 28, …, 228, 238, 255.  Here, the first step is
8, the next ones are 10 until final one which is 17.

Assuming the mapping is linear causes wrong results.  For example,
colour #1C1C1C is at index 234 but the code incorrectly mapped it
to index 233 (which is 0x121212).

To further improve the approximation, the new code shifts values
to use rounding during division rather than truncating them.  This
leads #070707 to be approximated by #080808 rather than #000000.