Include bash completion in deb package
While getting the correct path to install in from bash-completion using
`pkg-config` would strictly speaking be the right thing to do, simply
hardcoding it is the maintainers' preference. There is no difference on
the end result though for now, and unlikely to be in the future either.

Install the script as `steps.strip.outputs.BIN_NAME` as the filename
needs to match the called executable for bash-completion's on demand
loading to work. This `BIN_NAME` needs to stay in sync with
`env.PROJECT_EXECUTABLE` which is replaced in the template, but not
available for the deb build at this stage.

Closes https://github.com/sharkdp/bat/issues/1733