CICD: Significantly speed up syntaxes and themes regression testing
We run `cargo test` several times over in the matrix job, and there is no need
to run it again with updated assets.

To be able to skip buidling the dozens of extra dependencies required for
testing, we also convert the "all themes are present" test to a bash script
instead.