Changelog enriched with `cargo changelog` (a binary of `cargo-smart-release`)
This is the result of running `cargo changelog --write`, which
regenerates the changelog to amend information from the git history
while leaving user edits in tact. This works here as the changelog
format is already close to the one suggested by keepachanglog.

There are two auto-generated sections, `Commit Details` and `Commit
Statistics`. Both can be disabled by running `cargo changelog --without
commit-statistics commit-details`. The added header can just be replaced
with an empty line and won't be regenerated.

If `cargo smart-release` is used for release, a `cargo release`
compatible tag will be generated which carries the relevant section
from the changelog, as well as a GitHub release, if the `gh` executable
is found. Note that it generates special versions for each target to
get the best possible rendering.

Lastly, even without `cargo smart-release`, `cargo changelog` will
none-destructively regenerate the changelog and respect edits.

Please feel free to close this PR as it's definitely unsolicited, but
I thought I'd try in case the change is likable or can be made that way.