add zbat wrapper to decompress files automatically
This relies on bsdcat from archive-tools, which supports more formats
than just gzip or, failing that, falls back to gunzip -c with a
warning.

Previously[1], @sharkdp stated that bat would not support compressed
files, but would be open in implementing a wrapper. There was another
implementation proposed elsewhere[2], but it seems overengineered to
me.

[1] https://github.com/sharkdp/bat/issues/642#issuecomment-526816644
[2] https://github.com/sharkdp/bat/issues/237#issuecomment-617079288